Move buffers into the files that use them
