--- src/Mathematics.NET/Core/Buffers/RMTensor4Buffer4.cs
@@ -1,48 +1,0 @@
-﻿// <copyright file="RMTensor4Buffer4.cs" company="Mathematics.NET">
-// Mathematics.NET
-// https://github.com/HamletTanyavong/Mathematics.NET
-//
-// MIT License
-//
-// Copyright (c) 2023-present Hamlet Tanyavong
-//
-// Permission is hereby granted, free of charge, to any person obtaining a copy
-// of this software and associated documentation files (the "Software"), to deal
-// in the Software without restriction, including without limitation the rights
-// to use, copy, modify, merge, publish, distribute, sublicense, and/or sell
-// copies of the Software, and to permit persons to whom the Software is
-// furnished to do so, subject to the following conditions:
-//
-// The above copyright notice and this permission notice shall be included in all
-// copies or substantial portions of the Software.
-//
-// THE SOFTWARE IS PROVIDED "AS IS", WITHOUT WARRANTY OF ANY KIND, EXPRESS OR
-// IMPLIED, INCLUDING BUT NOT LIMITED TO THE WARRANTIES OF MERCHANTABILITY,
-// FITNESS FOR A PARTICULAR PURPOSE AND NONINFRINGEMENT. IN NO EVENT SHALL THE
-// AUTHORS OR COPYRIGHT HOLDERS BE LIABLE FOR ANY CLAIM, DAMAGES OR OTHER
-// LIABILITY, WHETHER IN AN ACTION OF CONTRACT, TORT OR OTHERWISE, ARISING FROM,
-// OUT OF OR IN CONNECTION WITH THE SOFTWARE OR THE USE OR OTHER DEALINGS IN THE
-// SOFTWARE.
-// </copyright>
-
-#pragma warning disable IDE0051
-
-using System.Runtime.CompilerServices;
-using Mathematics.NET.AutoDiff;
-using Mathematics.NET.DifferentialGeometry;
-using Mathematics.NET.DifferentialGeometry.Abstractions;
-
-namespace Mathematics.NET.Core.Buffers;
-
-/// <summary>Represents a buffer of 4 AutoDiffTensor4 delegates.</summary>
-/// <typeparam name="TT">A type that implements <see cref="ITape{T}"/>.</typeparam>
-/// <typeparam name="TN">A type that implements <see cref="IComplex{T}"/> and <see cref="IDifferentiableFunctions{T}"/>.</typeparam>
-/// <typeparam name="TPI">The index of the point on the manifold.</typeparam>
-[InlineArray(4)]
-internal struct RMTensor4Buffer4<TT, TN, TPI>
-    where TT : ITape<TN>
-    where TN : IComplex<TN>, IDifferentiableFunctions<TN>
-    where TPI : IIndex
-{
-    private Func<TT, AutoDiffTensor4<TN, TPI>, Variable<TN>> _element0;
-}
